feat: base and emtion dialogs
